feat(example-backend): run admin scripts via CLI (bun run script)
