+++ example-backend/package.json
@@ -61,6 +61,7 @@
     "lint": "biome check .",
     "lint:fix": "biome check --write .",
     "lint:unsafefix": "biome check --write --unsafe .",
+    "script": "bun run src/scripts/runAdminScript.ts",
     "seed": "bun run src/scripts/seed-test-data.ts",
     "start": "bun run src/index.ts",
     "start:prod": "bun run dist/index.js",
